+++ bin/ecbundle
@@ -9,13 +9,14 @@ usage()
     echo "Usage: "
     echo "       ecbundle --version"
     echo "       ecbundle create [--help]"
+    echo "       ecbundle populate  [--help]"
     echo "       ecbundle build  [--help]"
 }
 
 
 # Check at least one argument was provided:
 if [[ $# -eq 0 ]]; then
-  echo "ERROR: At least one argument expected: '--version', 'build' or 'create'"
+  echo "ERROR: At least one argument expected: '--version', 'build', 'create' or 'populate'"
   usage
   exit 1
 fi
@@ -33,8 +34,11 @@ elif [[ "build" == "$1"* ]]; then
 elif [[ "create" == "$1"* ]]; then
   shift
   ${SCRIPT_DIR}/ecbundle-create "$@"
+elif [[ "populate" == "$1"* ]]; then
+  shift
+  ${SCRIPT_DIR}/ecbundle-populate "$@"
 else
-  echo "ERROR: Expected 'build' or 'create' as first argument"
+  echo "ERROR: Expected 'build' or 'create' or 'populate' as first argument"
   usage
   exit 1
 fi
